Move AcctMgrInfo and AcctMgrRPCReply into AccountManager.kt.
Also rename AccountManager.kt to AccountManagerClasses.kt for clarity.
